+++ src/models/option.ts
@@ -1,8 +1,10 @@
-export class UniOption {
-    value;
-    label;
-    sublabel?;
-    obj?;
+import { UniListItem } from './list-item';
+export class UniOption implements UniListItem {
+    value?: string;
+    label: string;
+    sublabel?: string;
+    obj?: any;
+    icon?: string;
     constructor(obj: Object) {
         Object.assign(this, obj);
     }